Python/requirements.txt
Django
django-cors-headers
djangorestframework
psycopg2-binary
sqlparse    